Get cargo check passing + remove web components
