+++ Cargo.toml
@@ -15,7 +15,6 @@ members = [
    "fuel-indexer-schema",
    "fuel-indexer",
    "fuel-indexer-tests/components/indices/fuel-indexer-test",
-   "fuel-indexer-tests/components/web",
    "fuel-indexer-tests",
 ]
