+++ fuel-indexer-tests/assets/wallet.json
