--- fuel-indexer-tests/components/web/README.md
@@ -1,13 +1,0 @@
-# fuel-indexer-tests/components/web
-
-Web components used for testing
-
-### Usage
-
-```bash
-# Start the processes
-bash scripts/start_test_components.bash
-
-# Kill the processes
-bash scripts/kill_test_components.bash
-```
